Create functions that hold the route handling logic (no touching of existing api routes) (#7334)
## Describe your changes
This refactors the logic and some other structs and utils for the 3
document upload endpoints (index, upload, and process) into methods that
can be tested.

I went back and forth on this - it's not a perfect contract since we
aren't testing the API route itself, and so if someone just swapped our
what methods are called in the route handlers, no rust tests would fail
(or added logic around the refactored functions). So on one hand, the
best sort of test for this would be python tests

OTOH - we have a _ton_ of complex logic. Business logic around which
document_types go to `index`, whether or not to create decision_intents
if it's the appclip vs. not in `upload`, and certainly not least the
incode state machine in `process`. And these are all complicated by the
fact that we have 7 `IdDocKinds`, sandbox vs. prod etc. So, I feel that
refactoring the logic into unit/rust integration testable methods is the
best of the worlds.

This PR just pulls out the logic from the existing routes into separate
functions and files, it doesn't actually touch the prod routes.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
we're getting there

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->