[FP-6617] Simply create all fingerprints when data is added to vault (#7367)
## Describe your changes
After a lot of digging, I have concluded it is safe to get rid of some
convoluted logic on when to fingerprint data.

For background, we have two types of fingerprints:
- Global: only a handful of DIs are globally fingerprinted: email,
phone, ssn, and ein. They are only used programatically in order to
identify users that have a given email or phone during our login
identify flow. The SSN and EIN global fingerprints are not used at all.
- Tenant-scoped: we fingerprint more kinds of data, but salting with the
tenant ID such that knowledge of fingerprint values at one tenant
doesn't give knowledge of fingerprint values at another tenant.
Admittedly, I still only have medium conviction this is even necessary
for the volume of data that we are fingerprinting

Some data is globally fingerprinted, some data is globally AND
tenant-scoped fingerprinted, and other data (usually lower entropy/not
searchable) is not fingerprinted at all.

Now, given this context, our previous convoluted logic would:
1. Create globally-scoped fingerprints for data added during bifrost,
and then add tenant-scoped fingerprints when the workflow finished. The
initial idea behind delaying tenant-scoped fingerprints was that we
didn't want to allow searching in the dashboard on un-authorized fields.
But now, we auto-authorize most onboardings. And if we want to still
have this behavior for the few one-click users that abandon before
finshing, i think we should implement this a different way by including
whether data is authorized on the DL
2. Create tenant-scoped fingerprints for data added via tenant API. If
the user ends up later becoming portablized, we then create
global-scoped fingerprints when the workflow is finished. The initial
idea behind delaying creation of global-scoped fingerprints is that they
we didn't want a vault created via tenant API to be discoverable by
another tenant via the identify flow. But, in #7353, we have much more
context-aware logic to select a user matching a globally-scoped
fingerprint, and we simply won't match for a user that hasn't
OTP-verified their contact info.
3. (more modern) create both global and tenant-scoped fingerprints for
prefill data

As we've added more codepaths by which data can be added to a vault
(prefill, tenant-facing API, user-facing API), it has gotten extremely
convoluted to manage this conditional fingerprinting, so much so that at
the end of a workflow, we just try to make all global and tenant-scoped
fingerprints as a catch-all.

Given that the cases for delaying fingerprinting above are pretty much
obsolete, i think it is not worth keeping these convoluted codepaths and
the complexity to understand what types of fingerprints exist in a
vault.

## Caveats
The logic in `vw.create_authorized_fingerprints` is now obsolete for new
vaults, but is still needed for legacy vaults. I will run a backfill of
existing vaults to create all the fingerprints that should exist for all
pieces of data

## Test Plan
Even worse, I don't think we really had test coverage of this complex
logic. Let's see what integration tests fail...