Request more info dialog fix for cmd k (#7435)
The dialog was not showing the SSN option for cmd+k case. It makes sense
to unify this component and show the same dialog in both cmd+k and
actions instead of duplicating the component. We had 2 instances of it,
which made it a bit more complicated to maintain / update

![CleanShot 2024-01-02 at 09 55
45@2x](https://github.com/onefootprint/monorepo/assets/12160059/0c6fa03e-95f3-407e-b0b6-897c62f9f86a)