[Dashboard] Upload source maps to sentry + update sentry (#7441)
## Describe your changes
While looking at these errors
(https://onefootprint.slack.com/archives/C05SN1UTZLY/p1704232444767959),
I noticed it's very hard to understand them, given we're not uploading
source maps.

In this PR, we're:

- Updating next sentry to the latest version
- Uploading the source maps during the build, when we're deploying an
app to Vercel

If this goes well, the idea is to replicate to the other projects as
well.

## Caveats
One of the concerns is that source maps could make our code visible to
the end users. Sentry has an option that prevents this though:
https://docs.sentry.io/platforms/javascript/guides/nextjs/manual-setup/#use-hidden-source-map

## Test Plan

- To test locally, just build the app running `yarn build`. This should
not show anything differently from before.
- To see the source maps upload, you can take a look here:
https://vercel.com/footprint/dashboard/AkjkJeaKrorvYFttNJYKX1cYKffV.
Noticed we're printing "📦 Uploading source maps to Sentry". Sentry
provides an option to output all the uploaded files during build time,
but for now, we're not showing to keep the build logs cleaner

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->