New expandable nav for api docs (#7465)
elements in the api docs navigation are becoming longer and when
truncated are hard to differentiate. With this new nav behavior, when
the user hovers over the nav, we expand it slightly so it reveals more
of the content. When the user leaves the hover, we collapse it back to
leave more space for the content of the page.


https://github.com/onefootprint/monorepo/assets/12160059/3c7dca7c-8ae4-4b9e-9853-f3c103e2dfdf