[1/3] Create vault.is_identifiable column to replace is_portable (#7486)
## Describe your changes
The `vault.is_portable` column has been misnamed for a really long time
- right now, `is_portable` really just means "can we find this vault in
POST /hosted/identify"

Our metrics on airplane have always had an approximation of PIDs, and
it's becoming increasingly more important that our metrics reflect the
actual state of the world.

It's a little complex to actually query what a "PID" is - do we query
for portable data? Do we query for approved onboarding decisions made by
footprint? So, I'm instead going to denormalize "is this a PID" onto the
vault table and repurpose `vault.is_portable` to represent this exactly.

So, this adds a new column, `vault.is_identifiable`, whose value is the
same as the old `vault.is_portable`. In subsequent PRs, we will backfill
it and move reads to this new column. For now, it isn't read at all

## Caveats
Once we stop reading `is_portable` for the `POST /hosted/identify` case,
we can start writing proper values to it - a vault will start as
`is_portable = false` and graduate to `is_portable = true` when we
portablize data

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->