[FP-6968] Clean up airplane stats dashboard (#7501)
## Describe your changes
These queries had different methods for determining PID count. After the
work done in the stack starting with #7486, we actually have a
denormalized column that tells us exactly if a vault is portable. Now, a
PID is literally just a user that has portable data that we can prefill
during one-click onboardings.

The NYPID data also counted any vault that wasn't portable. Now, we only
count vaults created via API, which slightly decreases the NYPID count.
So, an NYPID is a vault made via API that isn't yet portable.
(technically, we might want to require an NYPID to also have
"portablizable" data, any `id.xxxx` DI. a vault with only a credit card
isn't really portablizable because they have no data that can be
portablized. but, we could always choose to portablize CCs in the future
if we want. just hard because of the aliases)

## Caveats
Not all PIDs and NYPIDs are created equal too - some PIDs/NYPIDs have
more data than others, and presumably the more full a PID is, the better

We use these metrics for investor updates and fundraising, so pls take a
critical look at these queries

## Test Plan
deployed to airplane

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->