[FP-6918] Add initial APIs to replace auth methods (#7431)
## Describe your changes
First attempt at a general API to replace an auth method in the vault.
Replacing an auth method requires (1) being signed in and having high
enough permissions to replace the CI and (2) proving that you own the
new auth method with a challenge.

So, this makes two APIS, `POST /hosted/user/challenge` and `POST
/hosted/user/challenge/verify` that generically allow you to initiate
any kind of challenge and verify it

## Caveats
This challenge code (login challenge, signup challenge, and now this new
replace auth method challenge) is getting very duplicative. It's
starting to smell like there's probably a cleaner abstraction we can
expose. Especially in the future - we will have a challenge to _add_
contact info without replacing it. Will noodle on this...

Still need to implement permissions that only allow certain auth methods
to be replaced based on what auth methods you used to log into your
vault - that's coming in #7471

And, in a future PR, will add more limitations on when these APIs can be
used

## Test Plan
Integration tests
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->