[FP-6935][FP-6950] Require token with correct scopes to update auth methods (#7446)
## Describe your changes
Will land this alongside the PR above it so the above PR is never
deployed without this. Just separating into separate PR for review.

The idea is we want pretty strict conditions to be met in order to
update a user's auth methods since we want to reduce the risk of ATOs as
we slowly roll out this feature
- The tenant must send a link to the user that allows them to update
their contact info - so there's friction in updating contact info in
that a human at the tenant must send a physcial link. This constraint
will likely be relaxed in the future, but it adds some safety in the
very beginning. For now, this means that an auth token created via
bifrost or via fp auth cannot be used to update auth methods
- The tenant should not be able to update their own user's auth methods
_without_ first verifying the existing auth methods. In short, inherited
auth cannot be used - auth must be made explicitly to Footprint by the
user before they update their contact info

## Caveats
These boolean flags we check are kind of "auth scopes." I'm thinking to
migrate them to be closer to auth scopes in the future

## Test Plan
Integration tests