Add TenantBusinessInfo (#7510)
## Describe your changes

<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

Adds new `tenant_business_info` table that stores the official business
name/address/phone for a Tenant. As part of our reseller agreement with
Lexis, we need to forward this information in every single one of our
requests we make to them.

I opted to make a new table and new private/protected endpoints to
set/get this (next PR) because:

- While `tenant` has `name`, this isn't generally a proper legal name
for our Tenants and I didn't want to modify a bunch of `tenant.name`'s
just for this
- We also need to store address + phone and it felt nicer to make a new
table here rather than add to the ever-growing `tenant` table
- I don't think we need to expose this in the settings page to Tenants
quite yet and it's also not clear to me that we want Tenant's to be able
to arbitrarily change this. Presumably we have their business info
stipulated in our contracts with them (and possibly in contracts with
specific vendors) and we'll be forwarding this to folks like Lexis so
it's something we should probably oversee changes to (which shouldn't
happen often anyway).

I also did not use our DI's/vault data models here just for ease.
But very open to other ideas if folks feeling strongly here :)

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? -->

unit test

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->