Fix random bug in pin input (#7555)
This PR fixes a weird issue in Styled Components during development.
Sometimes, pin inputs defaults to a `width: 100%` instead of the
intended `width: 40px` from `styled(Input)`. It seems the default
`width: 100%` kicks in when the Input lacks an explicit `width` prop. To
avoid this, let's play it safe and pass the width prop directly instead
of relying on a style.