Replace client_token `scopes` with `scope` (#7556)
## Describe your changes
We're going to add a new `scope`, vault_card, which doesn't really make
sense in the context of a vec of permissions. vault and decyrp are the
only two perms that can happen together, so i just made a custom scope
for it

## Caveats
Added some metrics to see who is using the old serialization - we can
deprecate the old API when they stop using it.

Build will fail backcompat because we changed the serialization shown on
docs. But the integration tests show that the application logic is still
backwards compatible

## Test Plan
Kept integration tests the same to test backcompat. Will update
integration tests in a future PR