[FP-6981] Add vault_card client token scope as shorthand to vault a card w random alias (#7557)
## Describe your changes
Adds a backend util to autogenerate the card alias so select doesn't
have to. We return it in the fields when the token is created, which
hopefully helps. Maybe they'll want it on their client as well


https://onefootprint.slack.com/archives/C04HU5VKLQK/p1704757476777759?thread_ts=1704755194.552109&cid=C04HU5VKLQK

## Caveats
This is really just a shorthand so it's surprising it adds that much
utility for select

## Test Plan
Integration test