Only allow adding _first_ auth credential using `add` action (#7537)
## Describe your changes
Our initial version of this product won't allow adding _secondary_ forms
of auth methods - only replacing contact info or adding it for the first
time. So, this changes the meaning of the `Add` action kind to only
allow adding for the first time

## Caveats
Haven't implemented `Add` for phone/email - will do in a follow-up

This also doesn't assert that Replace does, in fact, replace an existing
form of the auth method. Not the biggest deal since replacing requires
more strict auth

## Test Plan
Integration tests

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->