Permit features based on OCR class (#7611)
## Describe your changes
Handle FetchOCRResponse classes/issuing state

Didn't go the whole enum route for classes and issuing states since it
seemed overkill at this time, but will if we start adding a lot more of
these.

GA is the majority of flexcar's volume (~60%) but will add some more
states

There's a bit of complexity since we have some DL licenses that are
permits but for commercial vehicles. I'm not sure if that means you are
OK for non-commercial, but not for commercial. I don't really know a lot
about this so I'll ask Flexcar.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->