Produce AffiliatedWithBrokerOrFinra risk signal (#7623)
## Describe your changes

<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

Start writing a new `AffiliatedWithBrokerOrFinra` risk signal when a
user selects that option during the investor profile question.
The risk signal is produced when we produce the usual kyc/aml risk
signals. (ie we don't produce or remove this risk signal in perfect sync
with IPK's being vaulted).
We also have currently require a vres_id/vendor_api on each RiskSignal
so this does the hacky thing of just creating this risk signal with
theses from the Kyc vendor we are writing risk signals for. (which also
means that in this current stepup, we wouldn't produce this risk signal
if the obc was skip_kyc).

So overall, this is def a quick hack just to be able to quickly write a
rule for Composer that hard fails when the user has finra affiliation. I
think both (1) non-vendor/vres specific risk signals and (2) user input
dependent risk signals are interesting challenges we'll probably need to
solve soon, but I think this quick thing for Composer maybe isn't the
right impetus? (1) we might run into soon for our v0 of Dupes stuff. (2)
might not be necessary if longterm we support writing rules on data and
not just risk signals

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->