Fix flaky test: test_org_frequent_notes (#7643)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
Since multiple runs of integration tests share the same DB, we can't
look at global state (the list of frequent notes for a tenant). In this
PR I check that created frequent notes are a subset of all frequent
notes. It's a worse test but if fixes the flake :/

It's kind of concerning that tests don't get isolated environments,
since you can easily write tests that pass only due to side effects of
previous test runs (or in this case, fail due to the same thing).
Additionally, it's concerning that CI runs directly in prod -- a bad CI
test could potentially drop or corrupt prod data.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->