[FP-7034][FP-6976] Create employee-only toolbox section of dashboard (#7694)
## Describe your changes
Airplane is being deprecated - it used to house some utils for
employees. The long lived utils we'll move into the dashboard.

This creates the first version of the employees-only toolbox, so far
with only two tools:
- Copy your dashboard auth token to the clipboard
- Clean up a vault with the provided email/phone

## Caveats
It would be nice to make it very easy to add new "tools" to this
toolbox. I made it somewhat reusable now

I even think it would be nice to be able to omit translations and just
put inline strings since only employees will ever be consuming these
pages. But let me know what you think :)

## Test Plan
Locally:


https://github.com/onefootprint/monorepo/assets/21066341/c13e0099-11bd-46d0-9702-816ecf994078