[FP-6653] Denormalize is_live column on audit_log table (#7797)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
We'll heavily rely on an index on this column and we've had to
denormalize it on the similar access_event table in the past.

Note that `is_live` is nullable -- some audit logs might affect things
like org settings which are shared for live and non-live views of audit
logs. We could potentially call those shared event "live", but then the
non-live view of audit logs would counter-intuitively exclude those
events which affect the non-live environment.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->