[FP-6653] Add framework for audit log models and implement a few event types (#7779)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
This code provides the ability to write an audit log to the DB and read
back details into the same type. This model provides infallible
serialization when writing to the DB but fallible deserialization from
the DB. This is because we're overloading a single schema with multiple
event types using nullable columns, and these columns may be null for a
row where they should be not null. This is a compromise to avoid having
dozens of tables, or many tables that each have a share of the audit
logs.

I waffled a bunch about whether to use generics, enums, or trait
objects. This is what I settled on for now, but there's a decent chance
another approach ends up being more ergonomic. That's easy to switch out
if we come to that.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->
Unit tests, and it's also tested in the real app in the next PR in the
stack.

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->