[FP-5344] Speed up string operations in vault proxy (#7810)
## Describe your changes
The old method would do a string operation for _every_ token. Assuming
the string replace is an O(n) operation, this was O(n^2). Usually,
algorithmic complexity so rarely has an impact on practical runtime.
But, in some cases, we are [seeing
20s](https://ui.honeycomb.io/footprint-2e/environments/prod/datasets/fpc-api/result/ynjytiaw8n1/trace/9C7u7etGLPF?fields[]=s_name&fields[]=s_serviceName&span=815a4fa849517129)
just for this in-memory string operation 😂

<img width="1129" alt="image"
src="https://github.com/onefootprint/monorepo/assets/21066341/f969d132-5a31-45c7-a5fa-76c59f9deabb">


The new method simply replaces tokens while iterating through the string
for an O(n) operation.

This extracts _minimal_ tokens from the body. For example, if you have
`{{ {{ hello }} }}`, the outer sets of brackets should not be considered
part of the token - we consider the smallest set of characters when
parsing the token.

Since the code to actually extract the set of minimal tokens is pretty
complex, I created one function, `process_tokens`, that is used both to
parse out the tokens and to replace the tokens with the detokenized,
decrypted values

## Caveats
This makes one logic change: before this, we would magically consider
that `{{} }}` isn't a token, but we _would_ consider that `{{ { }}` is a
token. This seems to just have been an implementation detail. Now, we
will consider both of these to be tokens.

## Test Plan
Integration tests

This computes the proxy body with both methods for now and logs when
they differ