Give different name to CI jobs for mobile (#7852)
## Describe your changes
Github branch protections are a horrible product - they basically just check if a GHA job with the given name has passed before allowing merging.

Since the frontend_ci and mobile_ci jobs had the same names, passing mobile CI would allow landing a frontend PR and vice versa.

To make things works, we have a practice of running a "skip" job when there are changes outside of the folder in question. So, we'd run a "skip" mobile job on any frontend PR, which would near immediately make the CI green for the frontend jobs

## Caveats
I would like to rename the jobs for frontend_ci, but this may be disruptive to PRs that are unmerged. Maybe i will do later

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->