Store user token purpose in session (#7845)
## Describe your changes
We used to store this `is_created_via_api`, which would provide the ability to branch on some functionality when an auth token is created by the tenant via API or dashboard.

But, we're now need _more_ context on what the initial purpose of the token was. So, this adds more color to the auth token and instead stores the reason why the token was initially created in each auth session.

In a future PR, we'll actually store some context in the auth token on what auth methods are allowed ot be updated using the token

## Caveats
I'm not convinced this is the best way to store this info - in future PRs, we have some behavior branch based on the purpose of the token, which makes it feel like that purpose should be more represented in scopes. But for now, this is probably the cleanest way to add this functionality.

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->