Generate multiple invoices concurrently (#7880)
## Describe your changes
These are contributing to a lot of the 5xx alerts. We'll soon move this
to a cron, but for now, let's just make this slightly faster. The
invoice generation is taking just more than 60s, so i figure this will
help by allowing us to make some progress on generating tenant B's
invoice while waiting for some async operation to generate tenant A's
invoice


https://onefootprint.slack.com/archives/C03DQ288QQM/p1706290923059839?thread_ts=1706155619.769349&cid=C03DQ288QQM

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->